+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/metadata/resource/ShardingSphereResource.java
@@ -52,6 +52,16 @@ public final class ShardingSphereResource {
     }
     
     /**
+     * Get not existed resource name.
+     * 
+     * @param resourceNames resource names to be judged
+     * @return not existed resource names
+     */
+    public Collection<String> getNotExistedResources(final Collection<String> resourceNames) {
+        return resourceNames.stream().filter(each -> !dataSources.containsKey(each)).collect(Collectors.toSet());
+    }
+    
+    /**
      * Close data sources.
      * 
      * @param dataSources data sources to be closed
