+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/main/java/org/apache/shardingsphere/proxy/backend/text/distsql/rdl/impl/RDLBackendHandler.java
@@ -20,8 +20,11 @@ package org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl;
 import org.apache.shardingsphere.governance.core.registry.config.event.rule.RuleConfigurationsAlteredSQLNotificationEvent;
 import org.apache.shardingsphere.infra.config.RuleConfiguration;
 import org.apache.shardingsphere.infra.config.scope.SchemaRuleConfiguration;
+import org.apache.shardingsphere.infra.distsql.RDLUpdater;
 import org.apache.shardingsphere.infra.eventbus.ShardingSphereEventBus;
 import org.apache.shardingsphere.infra.metadata.resource.ShardingSphereResource;
+import org.apache.shardingsphere.infra.spi.ShardingSphereServiceLoader;
+import org.apache.shardingsphere.infra.spi.typed.TypedSPIRegistry;
 import org.apache.shardingsphere.proxy.backend.communication.jdbc.connection.BackendConnection;
 import org.apache.shardingsphere.proxy.backend.context.ProxyContext;
 import org.apache.shardingsphere.proxy.backend.response.header.ResponseHeader;
@@ -32,6 +35,7 @@ import org.apache.shardingsphere.sql.parser.sql.common.statement.SQLStatement;
 import java.lang.reflect.ParameterizedType;
 import java.util.Collection;
 import java.util.Optional;
+import java.util.Properties;
 import java.util.stream.Collectors;
 
 /**
@@ -42,38 +46,43 @@ import java.util.stream.Collectors;
  */
 public abstract class RDLBackendHandler<T extends SQLStatement, R extends SchemaRuleConfiguration> extends SchemaRequiredBackendHandler<T> {
     
+    static {
+        ShardingSphereServiceLoader.register(RDLUpdater.class);
+    }
+    
     public RDLBackendHandler(final T sqlStatement, final BackendConnection backendConnection) {
         super(sqlStatement, backendConnection);
     }
     
-    @SuppressWarnings("unchecked")
+    @SuppressWarnings({"rawtypes", "unchecked"})
     @Override
     protected final ResponseHeader execute(final String schemaName, final T sqlStatement) {
-        Class<R> configRuleClass = (Class<R>) ((ParameterizedType) getClass().getGenericSuperclass()).getActualTypeArguments()[1];
-        R currentRuleConfig = findCurrentRuleConfiguration(schemaName, configRuleClass).orElse(null);
-        check(schemaName, sqlStatement, currentRuleConfig);
-        doExecute(schemaName, sqlStatement, currentRuleConfig);
-        postChange(schemaName);
+        R currentRuleConfig = findCurrentRuleConfiguration(schemaName).orElse(null);
+        RDLUpdater rdlUpdater = TypedSPIRegistry.getRegisteredService(RDLUpdater.class, sqlStatement.getClass().getCanonicalName(), new Properties());
+        rdlUpdater.checkSQLStatement(schemaName, sqlStatement, currentRuleConfig, ProxyContext.getInstance().getMetaData(schemaName).getResource());
+        rdlUpdater.updateCurrentRuleConfiguration(schemaName, sqlStatement, currentRuleConfig);
+        postRuleConfigurationChange(schemaName);
         return new UpdateResponseHeader(sqlStatement);
     }
     
-    protected abstract void check(String schemaName, T sqlStatement, R currentRuleConfig);
-    
-    protected abstract void doExecute(String schemaName, T sqlStatement, R currentRuleConfig);
-    
-    private void postChange(final String schemaName) {
-        ShardingSphereEventBus.getInstance().post(
-                new RuleConfigurationsAlteredSQLNotificationEvent(schemaName, ProxyContext.getInstance().getMetaData(schemaName).getRuleMetaData().getConfigurations()));
-    }
-    
     @SuppressWarnings("unchecked")
-    private Optional<R> findCurrentRuleConfiguration(final String schemaName, final Class<R> configRuleClass) {
+    private Optional<R> findCurrentRuleConfiguration(final String schemaName) {
+        Class<R> ruleConfigClass = (Class<R>) ((ParameterizedType) getClass().getGenericSuperclass()).getActualTypeArguments()[1];
         for (RuleConfiguration each : ProxyContext.getInstance().getMetaData(schemaName).getRuleMetaData().getConfigurations()) {
-            if (configRuleClass.isAssignableFrom(each.getClass())) {
+            if (ruleConfigClass.isAssignableFrom(each.getClass())) {
                 return Optional.of((R) each);
             }
         }
         return Optional.empty();
+    }
+    
+    protected abstract void checkSQLStatement(String schemaName, T sqlStatement, R currentRuleConfig);
+    
+    protected abstract void updateCurrentRuleConfiguration(String schemaName, T sqlStatement, R currentRuleConfig);
+    
+    private void postRuleConfigurationChange(final String schemaName) {
+        ShardingSphereEventBus.getInstance().post(
+                new RuleConfigurationsAlteredSQLNotificationEvent(schemaName, ProxyContext.getInstance().getMetaData(schemaName).getRuleMetaData().getConfigurations()));
     }
     
     protected final Collection<String> getNotExistedResources(final String schemaName, final Collection<String> resourceNames) {
@@ -82,6 +91,6 @@ public abstract class RDLBackendHandler<T extends SQLStatement, R extends Schema
     
     private boolean isExistedResource(final String schemaName, final String resourceName) {
         ShardingSphereResource resource = ProxyContext.getInstance().getMetaData(schemaName).getResource();
-        return null != resource && resource.getDataSources().containsKey(resourceName);
+        return resource.getDataSources().containsKey(resourceName);
     }
 }
